pyproject.toml
[project]
name = "customer-support-ai-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "gradio>=5.0.0",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-ollama>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-weaviate>=0.1.0",
    "pydantic>=2.12.5",
    "pypdf>=6.6.0",
    "python-multipart>=0.0.9",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
    "weaviate-client>=4.9.0",
]
